requirements.txt
langchain>=0.1.14
langchain-community>=0.0.30
langchain-ollama>=0.0.6
python-dotenv>=1.0.1
